Fixes for Go code generation.
Timestamps have to be int64 (duh), and have to include json:"" around the field name for it to actually work (the documentation is terrible on this).

+++ src/main/java/net/socialgamer/cah/Constants.java
@@ -257,7 +257,7 @@ public class Constants {
     @GoDataType("[]CardSetData")
     CARD_SETS("css"),
     CLIENT_NAME("cn"),
-    @GoDataType("int")
+    @GoDataType("int64")
     CONNECTED_AT("ca"),
     @GoDataType("bool")
     ERROR("e"),
@@ -277,7 +277,7 @@ public class Constants {
     HAND("h"),
     @DuplicationAllowed
     ID_CODE(AjaxRequest.ID_CODE),
-    @GoDataType("int")
+    @GoDataType("int64")
     IDLE("idl"),
     IP_ADDRESS("IP"),
     /**
@@ -554,7 +554,7 @@ public class Constants {
      */
     @DuplicationAllowed
     SIGIL(AjaxResponse.SIGIL),
-    @GoDataType("int")
+    @GoDataType("int64")
     TIMESTAMP("ts"),
     @DuplicationAllowed
     @GoDataType("bool")
